requirement.txt
absl-py==0.13.0
aiohttp==3.9.0
aiosignal==1.2.0
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
autograd==1.6.2
autograd-gamma==0.5.0
biopython==1.82
blinker==1.6.2
boto3==1.34.10
botocore==1.34.10
Bottleneck==1.3.5
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
chex==0.1.85
click==8.1.7
cloudpickle==2.2.1
cryptography==41.0.7
cycler==0.12.1
cytoolz==0.12.2
dask==2023.11.0
dgl==0.9.1.post1
docstring-parser==0.15
exceptiongroup==1.2.0
filelock==3.13.1
flax==0.3.4
fonttools==4.25.0
formulaic==1.0.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==0.4.1
graphlib-backport==1.0.3
grpcio==1.48.2
huggingface-hub==0.20.3
hyperopt==0.2.7
idna==3.4
imagecodecs==2021.8.26
imageio==2.9.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
interface-meta==1.3.0
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.4
lifelines==0.27.8
lmdb==1.4.1
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.5.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
msgpack==1.0.7
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.8
packaging==23.2
pandas==1.2.4
partd==1.4.1
Pillow==10.0.1
pip==23.3.1
platformdirs==3.10.0
pluggy==1.4.0
pooch==1.7.0
prefetch-generator==1.0.3
protobuf==3.20.3
psutil==5.9.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==8.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
rdkit-pypi==2022.9.5
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-image==0.18.3
scikit-learn==1.3.2
scipy==1.10.1
setuptools==68.2.2
six==1.16.0
tape-proteins==0.5
tensorboard==2.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2021.7.2
timm==0.9.12
tokenizers==0.15.1
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torch-cluster==1.6.0+pt112cu113
torch_geometric==2.4.0
torch-scatter==2.0.9
torch-sparse==0.6.16+pt112cu113
torch-spline-conv==1.2.1+pt112cu113
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.37.2
typed-argument-parser==1.9.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
Werkzeug==2.2.3
wheel==0.41.2
wrapt==1.14.1
yarl==1.9.3
zipp==3.17.0
